added mobile styling for UX changes
