+++ fireflight/src/styles/App.scss
@@ -5,8 +5,6 @@
 @import 'vars';
 
 // Components
-@import 'navigation';
-@import 'navigation-profile';
 @import 'login';
 @import 'register';
 @import 'auth-forms';
@@ -154,7 +152,7 @@ button {
       height: 18px;
       margin-left: 21px;
       transition: 0.2s;
-    }  
+    }
   }
 }
