+++ fireflight/src/styles/_forms.scss
@@ -1,18 +1,15 @@
 // Map Form Container
 .map-form-container {
   position: absolute;
-  left: 50%;
-  right: 50%;
   display: flex;
   justify-content: center;
 
   @media (max-device-width: $mq-mobile) {
-    padding-left: 1.5rem;
   }
 
   //map form text
   .map-form-text {
-    color: white;
+    color: black;
     font-size: 1.4rem;
 
     @media (max-device-width: $mq-tablet) {
@@ -27,24 +24,24 @@
     }
   }
 
-  .signup-form-text {
-    color: white;
-    font-size: 1.6rem;
-    line-height: 2.2rem;
-    font-style: normal;
-    font-weight: normal;
-    position: absolute;
-    top: 57.9rem;
-    left: 6.6rem;
-    width: 40rem;
+  // // .signup-form-text {
+  // //   color: white;
+  // //   font-size: 1.6rem;
+  // //   line-height: 2.2rem;
+  // //   font-style: normal;
+  // //   font-weight: normal;
+  // //   position: absolute;
+  // //   top: 57.9rem;
+  // //   left: 6.6rem;
+  // //   width: 40rem;
 
-    @media (max-device-width: $mq-mobile) {
-      font-size: 1.3rem;
-      position: absolute;
-      top: 61.5rem;
-      left: 6rem;
-    }
-  }
+  // //   @media (max-device-width: $mq-mobile) {
+  // //     font-size: 1.3rem;
+  // //     position: absolute;
+  // //     top: 61.5rem;
+  // //     left: 6rem;
+  // //   }
+  // }
 
   // Address Input
   .address-input {
@@ -66,7 +63,7 @@
   }
 
   .react-geocoder {
-    width: 30rem;
+    width: 40rem;
     height: 5rem;
     background: #ffffff;
     border: 2px solid #9e9e9e;
@@ -77,7 +74,6 @@
     line-height: 2.7rem;
     display: flex;
     align-items: center;
-    //padding-left: 1.7rem;
     &::placeholder {
       font-size: 2rem;
       line-height: 2.7rem;
@@ -91,8 +87,9 @@
       line-height: 4.5rem;
     }
     input {
-      width: 28rem;
+      width: 35rem;
       margin-left: 0.5rem;
+      margin-right: 3.5rem;
       font-size: 2rem;
       line-height: 2.7rem;
       display: flex;
@@ -100,17 +97,18 @@
     }
 
     @media (max-device-width: $mq-mobile) {
-      width: 22rem;
+      width: 35rem;
       height: 4rem;
       left: 2rem;
       top: 11.8rem;
+      padding-left: 0rem;
       background: #ffffff;
       border-radius: 3px;
 
       input {
-        width: 19rem;
+        width: 34rem;
         margin-left: 1rem;
-        font-size: 1.4rem;
+        font-size: 1.6rem;
       }
     }
   }
@@ -143,49 +141,49 @@
     }
   }
 
-  // Radius Input
-  .radius-input {
-    width: 8.4rem;
-    height: 5rem;
-    background: #ffffff;
-    border: 2px solid #9e9e9e;
-    border-radius: 5px;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 2rem;
-    line-height: 2.5rem;
-    display: flex;
-    align-items: center;
-    margin-left: 1rem;
-    padding-left: 1rem;
+  // // Radius Input
+  // .radius-input {
+  //   width: 8.4rem;
+  //   height: 5rem;
+  //   background: #ffffff;
+  //   border: 2px solid #9e9e9e;
+  //   border-radius: 5px;
+  //   font-style: normal;
+  //   font-weight: normal;
+  //   font-size: 2rem;
+  //   line-height: 2.5rem;
+  //   display: flex;
+  //   align-items: center;
+  //   margin-left: 1rem;
+  //   padding-left: 1rem;
 
-    @media (max-device-width: $mq-mobile) {
-      width: 5rem;
-      height: 4rem;
-      background: #ffffff;
-      border-radius: 3px;
-      margin-top: 0;
-    }
+  //   @media (max-device-width: $mq-mobile) {
+  //     width: 5rem;
+  //     height: 4rem;
+  //     background: #ffffff;
+  //     border-radius: 3px;
+  //     margin-top: 0;
+  //   }
 
-    &::placeholder {
-      font-family: 'Open Sans';
-      font-style: normal;
-      font-weight: normal;
-      font-size: 2rem;
-      line-height: 2.7rem;
-      display: flex;
-      align-items: center;
-      color: #a9a9a9;
-      padding-left: 1.7rem;
+  //   &::placeholder {
+  //     font-family: 'Open Sans';
+  //     font-style: normal;
+  //     font-weight: normal;
+  //     font-size: 2rem;
+  //     line-height: 2.7rem;
+  //     display: flex;
+  //     align-items: center;
+  //     color: #a9a9a9;
+  //     padding-left: 1.7rem;
 
-      @media (max-device-width: $mq-mobile) {
-        font-size: 1.6rem;
-        padding: 0;
-      }
-    }
+  //     @media (max-device-width: $mq-mobile) {
+  //       font-size: 1.6rem;
+  //       padding: 0;
+  //     }
+  //   }
 
-    &::-moz-placeholder {
-      line-height: 4.5rem;
-    }
-  }
+  //   &::-moz-placeholder {
+  //     line-height: 4.5rem;
+  //   }
+  // }
 }
