--- fireflight/src/styles/_navigation-profile.scss
@@ -1,55 +1,0 @@
-.profile-container {
-  height: 4rem;
-  margin-top: 2rem;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-
-  @media (max-device-width: $mq-tablet) {
-    max-width: 100%;
-    padding-left: 10rem;
-  }
-
-  @media (max-device-width: $mq-mobile) {
-    max-width: 500px;
-    padding-left: 0;
-  }
-
-  .profile-item {
-    color: black;
-    font-weight: 600;
-    font-size: 1.8rem;
-    line-height: 2.5rem;
-    font-style: normal;
-    font-weight: normal;
-    padding: 2rem;
-    width: 10rem;
-
-    @media (max-device-width: $mq-mobile) {
-      padding: 0.25rem;
-      text-align: center;
-      font-size: 1.7rem;
-    }
-
-    &:hover {
-      background-color: white;
-    }
-  }
-
-  .profile-item.logout {
-    font-style: normal;
-    font-weight: normal;
-    font-size: 12px;
-    line-height: 16px;
-    background: #ffffff;
-    border: 0.5px solid #191a1a;
-    border-radius: 2px;
-    margin-left: 2rem;
-    padding: 1rem;
-    display: flex;
-    justify-content: center;
-    align-self: flex-end;
-  }
-}
